wrap field quantities with type annotations
Instead of treating the field quantities as just matrices, we explicitly
give them types.  For example, quantities on cell edges would have the
type `Edge{Primal}` while those on the dual cell edges would have the
type `Edge{Dual}`.  This should make it easier to test the correctness
of the operators, e.g.  the curl operator takes values on the dual
nodes to values on primal edges, divergence composed with curl is zero,
etc.