remove ghost cells from most fields
The only place where ghosts cells are really needed seem to be the
vorticity fields.  These changes make it so that ghost cells are
specified by simply enlarging the dimensions of the dual nodes.  The
dimensions of all the other field quantities should be derived from the
dimensions of the dual cell grid.